pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-credit-scoring"
version = "0.1.0"
description = "AI credit scoring and document query utilities"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "llama-index-core>=0.14.0",
    "llama-index-llms-openai",
    "llama-index-llms-ollama",
    "llama-index-embeddings-openai",
    "llama-index-embeddings-huggingface",
    "llama-index-vector-stores-faiss",
    "llama-index-vector-stores-chroma",
    "openai",
    "tiktoken",
    "ollama==0.6.1",
    "chromadb",
    "faiss-cpu==1.8.0",
    "PyPDF2==3.0.1",
    "python-docx==1.1.0",
    "openpyxl==3.1.2",
    "python-dotenv==1.0.0",
    "pandas==2.1.4",
    "numpy>=1.26.4,<2.0",
    "requests==2.32.3",
    "httpx==0.27.2",
    "fastapi==0.108.0",
    "uvicorn==0.25.0",
    "pydantic>=2.9,<3",
    "sqlalchemy>=2.0,<3",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "hypothesis==6.92.1",
]

[tool.setuptools.packages.find]
include = ["src*", "config*"]
